+++ src/modules/ArrayToDo.js
@@ -1,0 +1,3 @@
+const listObj = JSON.parse(localStorage.getItem('data-lis')) || [];
+
+export default listObj;
