+++ src/modules/savaDataList.js
@@ -1,0 +1,7 @@
+import listObj from './ArrayToDo.js';
+
+const saveData = () => {
+  localStorage.setItem('data-lis', JSON.stringify(listObj));
+};
+
+export default saveData;
