+++ src/modules/trashpin.png
